Added scrollview UI
